chore(patina): make the script-rule registry growable
The builtin script-rule registry data lived entirely inside
`linter/script_rules.rs`, which at 757 lines is already over the
350-line source-length guard. Because adding a script rule meant
appending to the entries array, the `RULE_*` name consts, and the
preset name lists -- all in that one over-limit file -- the file would
grow and the guard would reject the change.

Extract the registry data into a new growable `script_rules/registry/`
module, split data-only by concern so each file stays well under the
limit with room for future rules:

- `registry/mod.rs` (75): `BuiltinScriptRuleEntry` + `meta()`,
  `BuiltinScriptRuleMeta`, the preset consts, and re-exports.
- `registry/names.rs` (89): the `RULE_*` name consts and the ordered
  `ALL_BUILTIN_SCRIPT_RULE_NAMES` / `OPT_IN_SCRIPT_RULE_NAMES` lists.
- `registry/rules.rs` (244): the rule-instance static and the ordered
  `BUILTIN_SCRIPT_RULES` table.

`BUILTIN_SCRIPT_RULES` is now a `&'static [BuiltinScriptRuleEntry]`
slice instead of a fixed `[_; 25]` array, so no count constant has to
be bumped when a rule is added. `builtin_script_rules()` returns a
`Vec<BuiltinScriptRuleMeta>` and the name->entry lookup iterates the
slice by `rule_name`, removing the hand-maintained index match. Adding
a script rule now touches only the small `names.rs` + `rules.rs` files.

`script_rules.rs` shrinks 757 -> 400 (still over the limit, but the
guard only rejects over-limit files that grow). Behavior-preserving:
same 25 rules, same names, same presets, same diagnostic ordering;
`cargo test -p vize_patina` passes (1223 tests, 0 failed) with no
snapshot changes.

Refs #1629
